classification/requirements.txt
pokemontcgsdk
numpy
pandas
scikit-learn
matplotlib
seaborn
pymupdf
pillow
opencv-python
streamlit
streamlit-extras
pytesseract
pillow